fix: definitive fix for bug that wasnt allowing reb to deactivate when countdown was running
