+++ reb/inc/pins.h
@@ -1,0 +1,9 @@
+#ifndef H_PINS
+#define H_PINS
+
+// pin 0 - REB Countdown pin             Pino para indicar se o countdown do reb está ativo 
+
+#define REB_COUNTDOWN_PIN 0
+
+
+#endif
